Move code from index.ios.js into App component
